[api] move model MaintenanceStatistic => Statistics::MaintenanceStatistic
